Add clean timm compatibility for all pretrained models
Make any pretrained timm model work through the load/activation/benchmark
pipeline without hand-writing per-model configs.

- load_model: drop the hardcoded cluster sys.path + top-level RESMAX import
  that made the module unimportable off-cluster; import timm properly and
  load chresmax lazily with a clear error pointing at the Serre Lab fork.
- load_model: add a `transform: timm` sentinel that builds each model's
  native eval transform (input size, crop, interpolation, normalization)
  from its pretrained config, with no weight download.
- populate_timm: emit `transform: timm` per model so every pretrained timm
  model gets correct preprocessing automatically.
- benchmark: cast activations to float32 before fitting. They are stored as
  float16, and StandardScaler divides in-place in the input dtype, which
  overflows to inf for low-variance features. Affects all model sources.
- requirements: add timm (imported but previously unlisted).
- tests: cover module import without the fork and timm-native transform
  resolution (no GPU, no weight download).
- README: document timm benchmarking and the HMAX fork requirement.

+++ benchmark.py
@@ -43,8 +43,15 @@ def main(args):
             print(f"No activations found for layer: {layer}")
             continue
 
+        # Cast to float32: activations are stored as float16, and downstream
+        # StandardScaler divides in-place in the input dtype, which overflows
+        # to inf for low-variance features when left as float16.
         activations = (
-            activations.reshape(activations.shape[0], -1).detach().cpu().numpy()
+            activations.reshape(activations.shape[0], -1)
+            .detach()
+            .cpu()
+            .float()
+            .numpy()
         )  # [B, H * W * C]
         neural_responses, reliability = tvsd_dataset[: activations.shape[0]]
         reliability_mask = reliability > args.reliability_threshold
